% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MacroAlgaeIndicator_CumulativeCover.R
\name{MacroAlgaeIndicator}
\alias{MacroAlgaeIndicator}
\title{Macroalgae indicators}
\usage{
MacroAlgaeIndicator(indicator = c("CumulativeCover", "PropOpportunist",
  "NPerennials"), df, boundaries, depth_cutoff, std_depth = 7,
  std_haardsub = 50, n_iter = 10000, Indicator_pred_pct = 0.5)
}
\arguments{
\item{indicator}{Name of the indicator to be calculated. Possible values are
"CumulativeCover", "PropOpportunist", and "NPerennials"}

\item{df}{A dataframe with monitoring data for macroalgae according to the 
Danish National Marine Monitoring Program. The dataframe should contain the
following variables: 

\describe{ 
\item{kildestationnavn}{An identifier for the monitoring
transect name.}
\item{dato}{Date of the sampling.}
\item{dybde}{Depth of
the observation in meter.}
\item{hardbund_daekpct}{Percentage cover of
suitable substrate for macroalgae.}
\item{totcover_daekpct}{Total cover of
macroalgae relative to the suitable substrate.}
\item{proevetager}{Identifier of the person carrying out the monitoring
transect.}
\item{steneck}{The Steneck number of the species.}
\item{growth_strategy}{Either "P" for perennial, "O" for opportunist or "C"
for crustforming algae.} }}

\item{boundaries}{A vector of length 4 giving the status class boundary values
of the standardized indicator. The boundary values should be
given in the order high-good, good-moderate, moderate-poor, poor-bad.}

\item{depth_cutoff}{The cut-off depth for physical exposure. Observations 
whith depths shallower than this value are not used in the indicator
calculation.}

\item{std_depth}{The depth for which  the indication is calculated. Default
is 7 meters.}

\item{std_haardsub}{The suitable substrate percentage for which the indicator
if calculated. Default it 0.5.}

\item{n_iter}{Maximum number of iteration for Montecarlo simulation.}

\item{Indicator_pred_pct}{Probability for predicting the percentile in the simulated distribution. Default is 0.5 for the median.}
}
\value{
TODO
}
\description{
Macroalgae indicators
}
\examples{
data("alsfjord_2013_2016")
indicator <- MacroAlgaeIndicator("CumulativeCover", alsfjord_2013_2016, boundaries = c(74.4, 40.8, 20.4, 13.8), depth_cutoff = 1)
}
